Separate streams and sentry_flink libraries (#31)
The flink integration is really heavyweight for the development environment.
This represents a problem at this stage and it prevents us from moving ahead
in putting the API in front of users inside the sentry code base.

This is because:

Outside of python 3.11 you need java on the local environment to pip install pyflink (there are no wheels above python 3.11)
running the flink tests still require java
we do not have an image of sentry/snuba with the streamign api inside
So let's make it more lightweight.
Our api creates an abstraciton layer on top of a dataflow platform underneath.
We do not need to use Flink in the dev environment at all as others have
a semantics that is similar enough for our purposes.

This PR splits the streams python package into two: one is all the flink
agnostic stuff. This is enough to run the whole platform on Bytewax or others.
The second package depends on the first and contains all the Flink dependencies.

This means we can move on testing the api in the other code bases without bringing
over Flink while we figure out a way to re-package Flink for the dev environment