Adapte the runner to run the Arroyo adapter (#78)
The arroyo adapter was not wired onto the runner so we could not
actually run Arroyo consumers. This wires it up together.

Specifically:

the loader can instantiate an Arroyo adapter now.
There are types to fix but that requires a larger change.
Initialized the logger with a reasonable format and replaced
some print with log statements
Added a signal handler on the runner
Fixed the readme by adding a quickstart section