Allow re-initialization of the RustOperatorDelegate (#118)
The previous implementation of the Python Delegate to execute Python
Streaming Steps inside the Rust runtime was missing a way for the
rust code to reinitialize the Streaming Step upon reinitialization
of the Rust Arroyo Processing Strategy.

This is critical during rebalance as we clean up the state of the
processing strategy when partitions are assigned/revoked.

This PR introduce a RustOperatorFactory that builds the delegate.
Now we pass an instance of the factory to the runtime so the runtime
is able to initialize the delegate and the factory itself can
maintain state (like ProcessPools as we do on parallel processing).

This follows the Arroyo design as it is Arroyo that decides when to
instantiate the processing strategies.