Support Reduce in the Rust Adapter as a Python delegate (#119)
This uses the RustOperatorDelegate to make the Python Reduce step
working on the Rust Adapter.

We create the reduce step as an Arroyo Python strategy just like in the
Python adapter. Then we wire an OutputRetriever as following step.
This is used to colelct the result.
The above is then wrapped into the ReduceDelegate which extends
RustOperatorDelegate, which is thus manageable by the Rust runtime.