fix(flatmap): Change the FlatMap signature to match usage
The FlatMap takes a single message as input (i.e. a batch) and outputs 1->N messages. Currently have
the output type be a `Iterable[TOut]` and the next step would be required to iterate the output from
the FlatMap.

Note that this is a temporary measure: the API currently doesn't have the capability of handling
1->N style primitives.