feat(api): Allow the Pipeline to have a chain API
This changes the Pipeline class to have a chain-style API. See simple_map_filter.py for an example
of how this looks.

In order to make this work, the Pipeline needs to be aware of the previous step so it can be passed
in as the input to the new step. Also, instead of the steps being registered when they are
instantiated, they are added to the pipeline manually.