ref(rust): Use python for instance check when parsing error in PythonAdapter (#166)
In python adapter, we are casting the class name of the exception into a
string in order to perform type checking. This method is not precise and
prone to undesired behaviours as any exception with the same name will
be considered the correct type.

This PR uses the pyo3 `import_exception` macro to get the actual
exception from arroyo and use the python interpreters isinstance
function to check types.