feat(api): Allow the Pipeline to have a chain API (#162)
This changes the Pipeline class to have a chain-style API. See
simple_map_filter.py for an example
of how this looks.

Making this work required refactoring some of the existing code. Some of
the major changes:
- `Pipeline` is initialized with a Source or Branch directly, so that a
truly empty `Pipeline` can't exist.
- Remove the concept of multiple sources from the Pipeline (for now).
- Remove the `add` function from the Pipeline. If a Pipeline can only
have one source, adding another Pipeline with a different source doesn't
make sense.
- Steps are no longer added to a pipeline when they are initialized.
They have to be added to the
  pipeline using the `.apply()` method.
- There is now a `ComplexStep` class that represents views on top of
normal steps. This is a way to
  provide syntactic sugar for common patterns (Parser, Batch etc.)
- The adapters have the ability to directly handle the `ComplexStep`
class, otherwise the step will
  be transformed into a simple step before being adapted.
- The way that `Broadcast` and `Router` work was changed to align closer
to how Chain was implementing
them. The steps now take an array of other pipelines (with a `Branch` as
their start) and will
merge those branches into the main pipeline. The merging now doesn't
filter out the `Branch` steps
  (this was being done inconsistently).
- `Pipeline` also has a `root` attribute that can be used to get the
root of the graph. Used for finding the initial
  `Branch` steps for routers and broadcasts.
- Since the steps no longer are aware of their place in the graph, the
`DummyAdapter` class isn't
  able to build a full picture of the graph. It had to be simplified.
- While logically the `Chain` API remains unchanged, the names of the
functions were changed to prevent
  overlap with the underlying `Pipeline` functions
This is the first step of the proposals made here:
https://github.com/getsentry/streams/pull/152#discussion_r2159537084