feat(api): Remove the chains.py API (#173)
Now that the Pipeline class can directly achieve the same API as the
chains.py file, use that in all cases and remove the chains.py file as
tech debt.

Also add a `route`, `broadcast` and `segment` function to the pipeline
to emulate the API exactly.

Follow up to https://github.com/getsentry/streams/pull/162