feat: Barely functional version of Rust map and Rust filter (#177)
Make it possible to call Rust functions from pipeline steps. So far only
Map and Filter are supported.

In order to make a Rust function referencable from our pipeline DSL
(which is always Python), the user has to package their functions as a
Python package. We provide macros to do that for the user, and a basic
framework for converting input/output types.

For now we provide only one "builtin" way to convert types using
`convert_via_json!`. But the user can choose to implement
`FromPythonObject` etc themselves, and pick protobuf, or something
entirely handwritten using pyo3 APIs.

User is responsible for writing their own `.pyi` files to make the
pipeline typesafe. Ideally this would be automatically generated too,
but we can add it later. It also depends on how they choose to convert
types. We probably need additional features in our schemas repos
(sentry-kafka-schemas and sentry-protos) to provide a mapping from
generated Rust to generated Python types, so that users can't
accidentally pick the wrong pair of rust/python types.

The `rust_simple_map_filter` example should show the envisioned API
end-to-end.

Under the hood, we call the Rust function through the GIL, but release
the GIL ASAP. We have a few alternatives:

* We can do unsafe pointer casting. This was the initial version but i
don't think it is actually _sound_. Different versions of the Rust
compiler can generate different function calling conventions (in theory)
-- it does however completely eliminate the GIL.
* We can perform the function call in batches to amortize GIL overhead.
TBD as to whether this is actually needed for performance.
* We can use GRPC, but I think this just adds overhead and solves no
problems right now.

The macro hides all of this from the API. So we can implement these
improvements at a later point. The important thing is that we have a
stable API for now, and we can worry about performance later.