ref(typechecking): remove msg_type from instantiation arg (#196)
`msg_type` arg is currently only used for type checking, so the rest of
the steps know which type the `TransformFuncReturnType` is. This arg is
confusing because `msg_type` is a required arg for `Parser` and
`BatchParser` classes, making it sound like this arg is in charge of
schema validation. Both myself and Andrew were under the impression that
this was for schema validation purposes upon first encountering this
code

Type checking is now done by explicitly stating the type inside `Parser`
and `BatchParser` classes, see updated examples. This is more intuitive
for the application code, because it's typing things, being used for
type checking purposes.