ref(metrics): Add metrics for tracking pipeline status (#184)
This attempts to wrap steps with automatic metrics, so that users can
see information about the
input/output of their steps as well as duration timings. The metrics are
only implemented for the
rust arroyo runtime, as there is no easy way to generically add metrics
for all runtimes.

This also attempts to initialize the metrics for arroyo, so that arroyo
will log its metrics if that
run time is being used.