fix(python_adapter): more __yield_messages fixes (#202)
https://github.com/getsentry/streams/pull/203 fixed the initial error by
returning messages as a Sequence instead of a generator, but this caused
another issue where watermarks weren't handled by the python adapter's
`handle_py_return_value` function.

This required implementing `From Py<PyAny> for Watermark` to allow
converting the python `PyWatermark` returned from the `PythonAdapter`
into a rust `PyWatermark`.

A rust test for this code path is coming in a future PR.