ref: Random fixes from profiling SBC (#216)
## Pickling

A change to rust_arroyo broke the pickleability of functions, because
each function was wrapped in an unpickleable closure. So the SBC
consumer worked until somebody tried to enable multiprocessing.

Fix that, and add an assertion that all functions must be pickleable
regardless of how the deployment is configured. A function now needs to
be pickleable even if the runtime doesn't have multiprocessing.

Moved around a bunch of lambdas in tests into global functions as a
result of that.

## Validation

There were a few bugs with validation and deployment config overrides.

Specifically, it was not possible to init `Batch()` in the pipeline
definition, and pass the batch time and size parameters only in
deployment config.

Fix this by splitting out validation into a separate function,
`validate`, so the runtime has full control when validation occurs.
Validation should occur only after overrides have been applied, and it
should take overrides into account.

There were discussions about whether it even makes sense that the
runtime adapter has to deal with overrides. I didn't fix that yet, but
probably good to followup on that.

Also, the default batch time is now 10 seconds instead of None, i.e.
it's now actually valid to set up a batch without any explicit config
for batch time/size.

## Error logging

In `consumer.rs`, we swallow all errors coming from arroyo
streamprocessor. This made the above pickling issues entirely invisible.

## Release profile

We're adding more debuginfo to the release profile to make the
stacktrace prettier. There doesn't seem to be any overhead on throughput
in the sandbox, but it made the profiles much nicer. I imagine that it
will make the stacktraces in sentry nicer too.

## Random fixes

Updated Cargo.lock in one example because apparently it was severely out
of date with the rest. We keep having that kind of problem as
sentry_streams changes its own dependencies, which affects the deptree
of that example.